added densenet and made changes to convnet and basenetwork class
